connected sanity with react app
